[REFACTOR] Move encryption messaging structs to own module
