+++ components/butterfly/src/rumor/service_file.rs
@@ -11,7 +11,7 @@ use crate::{error::{Error,
             rumor::{Rumor,
                     RumorPayload,
                     RumorType}};
-use habitat_core::{crypto::keys::{box_key_pair::EncryptedSecret,
+use habitat_core::{crypto::keys::{EncryptedSecret,
                                   KeyCache},
                    service::ServiceGroup};
 use std::{borrow::Cow,
