+++ components/core/src/crypto/keys/box_key_pair.rs
@@ -12,7 +12,8 @@ use super::{super::{ANONYMOUS_BOX_FORMAT_VERSION,
             KeyPair,
             KeyRevision,
             KeyType};
-use crate::{crypto::keys::NamedRevision,
+
+use crate::{crypto::keys::encryption::EncryptedSecret,
             error::{Error,
                     Result}};
 use serde_derive::{Deserialize,
@@ -25,10 +26,8 @@ use sodiumoxide::crypto::{box_::{self,
                           sealedbox};
 use std::{borrow::Cow,
           convert::TryFrom,
-          fmt,
           path::Path,
-          str,
-          str::FromStr};
+          str};
 
 #[derive(Debug)]
 pub struct BoxSecret<'a> {
@@ -36,216 +35,6 @@ pub struct BoxSecret<'a> {
     pub ciphertext: Vec<u8>,
     pub receiver:   Option<&'a str>,
     pub nonce:      Option<Nonce>,
-}
-
-/// An encrypted message sent anonymously to a recipient. The message
-/// was encrypted with the recipient's public key, and can only be
-/// decoded by the recipient's private key. The identity of the sender
-/// cannot be known.
-pub struct AnonymousBox {
-    // TODO (CM): if it's really anonymous, then sender is really the recipient!!!!
-    /// The identity of the keypair that was used to encrypt (and
-    /// thus, must be used to decrypt) this message.
-    sender:     NamedRevision,
-    /// The encoded ciphertext of the message.
-    ciphertext: Vec<u8>,
-}
-
-impl AnonymousBox {
-    pub fn new(sender: NamedRevision, ciphertext: Vec<u8>) -> Self { Self { sender, ciphertext } }
-
-    pub fn sender(&self) -> &NamedRevision { &self.sender }
-
-    pub fn ciphertext(&self) -> &[u8] { &self.ciphertext }
-}
-
-impl fmt::Display for AnonymousBox {
-    /// Implements Habitat's String formatting for sending signed
-    /// anonymous encrypted messages. The version and sender are in
-    /// plaintext, while the ciphertext is base64 encoded.
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f,
-               "{}\n{}\n{}",
-               ANONYMOUS_BOX_FORMAT_VERSION,
-               self.sender,
-               base64::encode(&self.ciphertext))
-    }
-}
-
-/// An encrypted message signed by the sender (and thus traceable to
-/// that sender) intended for a receiver. It was encrypted using the
-/// sender's secret key and the recipient's private key, and can only
-/// be decrypted with the sender's public key and the recipient's
-/// secret key.
-///
-/// See `sodiumoxide::crypto::box_` for further details.
-pub struct SignedBox {
-    /// The identity of the keypair of the sender of this
-    /// message. Only the owner of this pair's secret key could have
-    /// sent this message, and its public key must be used to decrypt
-    /// it.
-    sender:     NamedRevision,
-    /// The identity of the keypair of the recipient of this
-    /// message. The pair's public key was used to encrypt this
-    /// message (thus "addressing" it to the receiver), and only the
-    /// owner of this pair's secret key man decrypt it.
-    receiver:   NamedRevision,
-    /// The encrypted ciphertext of the message
-    ciphertext: Vec<u8>,
-    /// The cryptographic nonce used to encrypt the message.
-    nonce:      Nonce,
-}
-
-impl SignedBox {
-    pub fn new(sender: NamedRevision,
-               receiver: NamedRevision,
-               ciphertext: Vec<u8>,
-               nonce: Nonce)
-               -> Self {
-        Self { sender,
-               receiver,
-               ciphertext,
-               nonce }
-    }
-
-    pub fn sender(&self) -> &NamedRevision { &self.sender }
-
-    pub fn receiver(&self) -> &NamedRevision { &self.receiver }
-
-    pub fn ciphertext(&self) -> &[u8] { &self.ciphertext }
-
-    pub fn nonce(&self) -> &Nonce { &self.nonce }
-}
-
-impl fmt::Display for SignedBox {
-    /// Implements Habitat's String formatting for sending signed
-    /// encrypted messages. The version, sender, and receiver are in
-    /// plaintext, while the nonce and ciphertext are base64 encoded.
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f,
-               "{}\n{}\n{}\n{}\n{}",
-               BOX_FORMAT_VERSION,
-               self.sender,
-               self.receiver,
-               base64::encode(self.nonce),
-               base64::encode(&self.ciphertext))
-    }
-}
-
-pub enum EncryptedSecret {
-    Anonymous(AnonymousBox),
-    Signed(SignedBox),
-}
-
-impl EncryptedSecret {
-    /// Unwraps this to an anonymous secret, if it is actually anonymous.
-    pub fn anonymous(self) -> Result<AnonymousBox> {
-        match self {
-            Self::Anonymous(anonymous) => Ok(anonymous),
-            _ => Err(Error::CryptoError("Not an anonymous secret!".to_string())),
-        }
-    }
-
-    /// Unwraps this to a signed secret, if it is actually signed.
-    pub fn signed(self) -> Result<SignedBox> {
-        match self {
-            Self::Signed(signed_box) => Ok(signed_box),
-            _ => Err(Error::CryptoError("Not an signed secret!".to_string())),
-        }
-    }
-
-    /// Helper function to parse an `EncryptedSecret` from raw bytes.
-    pub fn from_bytes<B>(bytes: B) -> Result<EncryptedSecret>
-        where B: AsRef<[u8]>
-    {
-        str::from_utf8(bytes.as_ref())?.parse()
-    }
-}
-
-/// Encapsulates parsing logic for all variants of `EncryptedSecret`,
-/// since they overlap a great deal.
-impl FromStr for EncryptedSecret {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self> {
-        let mut lines = s.lines();
-
-        let version =
-            lines.next()
-                 .ok_or_else(|| {
-                     Error::CryptoError("Corrupt payload, can't read version".to_string())
-                 })
-                 .map(|line| {
-                     match line {
-                         BOX_FORMAT_VERSION | ANONYMOUS_BOX_FORMAT_VERSION => Ok(line),
-                         _ => Err(Error::CryptoError(format!("Unsupported version: {}", line))),
-                     }
-                 })??;
-
-        let is_anonymous = version == ANONYMOUS_BOX_FORMAT_VERSION;
-
-        let sender =
-            lines.next()
-                 .ok_or_else(|| {
-                     Error::CryptoError("Corrupt payload, can't read sender key name".to_string())
-                 })?
-                 .parse()?;
-
-        let receiver = if is_anonymous {
-            None
-        } else {
-            let r = lines.next()
-                         .ok_or_else(|| {
-                             Error::CryptoError("Corrupt payload, can't read receiver key \
-                                                         name"
-                                                              .to_string())
-                         })?
-                         .parse()?;
-            Some(r)
-        };
-        let nonce = if is_anonymous {
-            None
-        } else {
-            let n =
-                lines.next()
-                     .ok_or_else(|| {
-                         Error::CryptoError("Corrupt payload, can't read nonce".to_string())
-                     })
-                     .map(base64::decode)?
-                     .map_err(|e| Error::CryptoError(format!("Can't decode nonce: {}", e)))
-                     .map(|bytes| Nonce::from_slice(bytes.as_ref()))?
-                     .ok_or_else(|| Error::CryptoError("Invalid size of nonce".to_string()))?;
-            Some(n)
-        };
-
-        let ciphertext =
-            lines.next()
-                 .ok_or_else(|| {
-                     Error::CryptoError("Corrupt payload, can't read ciphertext".to_string())
-                 })
-                 .map(base64::decode)?
-                 .map_err(|e| Error::CryptoError(format!("Can't decode ciphertext: {}", e)))?;
-
-        if is_anonymous {
-            Ok(EncryptedSecret::Anonymous(AnonymousBox { sender, ciphertext }))
-        } else {
-            Ok(EncryptedSecret::Signed(SignedBox { sender,
-                                                   receiver:
-                                                       receiver.unwrap(),
-                                                   ciphertext,
-                                                   nonce:
-                                                       nonce.unwrap() }))
-        }
-    }
-}
-
-impl<'a> From<EncryptedSecret> for WrappedSealedBox<'a> {
-    fn from(payload: EncryptedSecret) -> Self {
-        match payload {
-            EncryptedSecret::Anonymous(anon) => Self(Cow::Owned(anon.to_string())),
-            EncryptedSecret::Signed(signed) => Self(Cow::Owned(signed.to_string())),
-        }
-    }
 }
 
 ////////////////////////////////////////////////////////////////////////
@@ -336,6 +125,16 @@ impl<'a> WrappedSealedBox<'a> {
         line.ok_or_else(|| Error::CryptoError("Corrupt payload, can't read ciphertext".to_string()))
             .map(base64::decode)?
             .map_err(|e| Error::CryptoError(format!("Can't decode ciphertext: {}", e)))
+    }
+}
+
+// Temporary !!
+impl<'a> From<EncryptedSecret> for WrappedSealedBox<'a> {
+    fn from(payload: EncryptedSecret) -> Self {
+        match payload {
+            EncryptedSecret::Anonymous(anon) => Self(Cow::Owned(anon.to_string())),
+            EncryptedSecret::Signed(signed) => Self(Cow::Owned(signed.to_string())),
+        }
     }
 }
