+++ components/core/src/crypto/keys/encryption.rs
@@ -1,0 +1,5 @@
+mod message;
+
+pub use message::{AnonymousBox,
+                  EncryptedSecret,
+                  SignedBox};
