+++ components/core/src/crypto/keys/origin_key.rs
@@ -1,6 +1,6 @@
-use crate::{crypto::{keys::{box_key_pair::{AnonymousBox,
-                                           EncryptedSecret,
-                                           WrappedSealedBox},
+use crate::{crypto::{keys::{box_key_pair::WrappedSealedBox,
+                            encryption::{AnonymousBox,
+                                         EncryptedSecret},
                             Key,
                             KeyRevision,
                             NamedRevision},
