+++ components/core/src/crypto/keys/user_key.rs
@@ -1,6 +1,6 @@
-use crate::{crypto::{keys::{box_key_pair::{EncryptedSecret,
-                                           SignedBox,
-                                           WrappedSealedBox},
+use crate::{crypto::{keys::{box_key_pair::WrappedSealedBox,
+                            encryption::{EncryptedSecret,
+                                         SignedBox},
                             Key,
                             KeyRevision,
                             NamedRevision,
